emailsender/requirements.txt
asgiref==3.8.1
attrs==25.1.0
cffi==1.17.1
chardet==5.2.0
cryptography==44.0.1
dj-database-url==2.3.0
Django==5.1.6
django-extensions==3.2.3
django-flanker==0.3.0
django-tinymce==4.1.0
djangorestframework==3.15.2
flanker==0.9.11
idna==3.10
pillow==11.1.0
ply==3.11
psycopg2==2.9.10
pycparser==2.22
pycryptodome==3.21.0
pyngrok==7.2.3
python-decouple==3.8
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
six==1.17.0
sqlparse==0.5.3
tld==0.13
typing_extensions==4.12.2
tzdata==2025.1
WebOb==1.8.9